+++ Sources/CodexTokenCostApp/Support/WindowOpeningSupport.swift
@@ -3,18 +3,7 @@ import SwiftUI
 
 @MainActor
 enum WindowOpeningSupport {
-    static func showOrRevealMainWindow(openWindow: OpenWindowAction) {
-        NSApp.setActivationPolicy(.regular)
-        NSApp.activate(ignoringOtherApps: true)
-
-        if let window = NSApp.windows.first(where: { $0.identifier?.rawValue == "main" }) {
-            window.makeKeyAndOrderFront(nil)
-        } else {
-            openWindow(id: "main")
-        }
-    }
-
-    static func openSingletonWindow(id: String, openWindow: OpenWindowAction) {
+    static func openWindow(id: String, openWindow: OpenWindowAction) {
         NSApp.setActivationPolicy(.regular)
         NSApp.activate(ignoringOtherApps: true)
         openWindow(id: id)